Move all autowrap stuff into the same directory
